+++ CHANGELOG.md
@@ -53,4 +53,6 @@
 - Release new models
 
 ## 0.3.3
+
 - Improvement in the loading of data during prediction. We now use a data loader.
+- Fixed missing import in parser module